% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catania.R
\docType{data}
\name{Catania}
\alias{Catania}
\title{Pattern of sexual contacts in AMEN study}
\format{Four-by-four numeric matrix with dimnames.

\preformatted{
          woman
man        black hispanic white other
  black    0.258    0.016 0.035 0.013
  hispanic 0.012    0.157 0.058 0.019
  white    0.013    0.023 0.306 0.035
  other    0.005    0.007 0.024 0.016
}}
\source{
Newman, M. (2003) "Mixing patterns in networks" Arxiv:cond-mat/0209450 v2
}
\usage{
Catania
}
\description{
Contact layer of the mixing matrix of men and women in US based on "Aids in
Multi-Ethnic  Neighborhoods" (AMEN). Based on Newman (2003).
}
\examples{
data(Catania)

# assortativity
ep <- sum(Catania \%*\% Catania)
( sum(diag(Catania)) - ep ) / ( 1 - ep )
}
\references{
Catania et al. (1992) "The population-based AMEN (AIDS in Multi-Ethnic
Neighborhoods) study" American Journal of Public Health 82, 284-287

Morris, M. (1995) "Data driven network models for the spread of infectious
disease". In D. Mollison (ed.) "Epidemic Models: Their Structure and
Relation to Data", pp. 302-322, Cambridge University Press, Cambridge

Newman, M. (2003) "Mixing patterns in networks" Arxiv:cond-mat/0209450 v2
}
\keyword{datasets}
