% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classroom.R
\docType{data}
\name{Classroom}
\alias{Classroom}
\title{Classroom network}
\format{Object of class "igraph" with a directed network of size 26. Vertex
attribute \code{gender}, takes values "Boy" or "Girl".}
\usage{
Classroom
}
\description{
A directed network of ties "with whom do you like to play with?" collected in
a school class of 9-year olds.
}
\references{

}
\keyword{datasets}
