% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netseg-package.R
\docType{package}
\name{netseg-package}
\alias{netseg}
\alias{netseg-package}
\title{netseg: Measures of Network Segregation and Homophily}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implementations of measures of (social) network segregation and
	homophily. See Bojanowski & Corten (2014) <doi:10.1016/j.socnet.2014.04.01> for
	details.
}
\author{
\strong{Maintainer}: Michal Bojanowski \email{michal2992@gmail.com} (\href{https://orcid.org/0000-0001-7503-852X}{ORCID})

}
\keyword{internal}
