% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orwg.R
\name{orwg}
\alias{orwg}
\alias{orwg.table}
\alias{orwg.igraph}
\alias{orwg.default}
\title{Odds ratio of existence of within-group ties}
\usage{
orwg(object, ...)

\method{orwg}{table}(object, ...)

\method{orwg}{igraph}(object, vattr, ...)

\method{orwg}{default}(object, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{vattr}{character scalar or any vector, name of the vertex attribute or
the attribute itself (as a vector)}
}
\value{
Numeric value of the measure.
}
\description{
Odds ratio for connected, as opposed to disconnected, dyads depending
whether it is between- or within-group, i.e. how much more likely the dyad
will be connected if it is within-group.
}
\details{
The measure takes values, like all odds ratios, from (0; Inf).

Method for mixing matrices.

Method for igraphs
}
\examples{
orwg(WhiteKinship, "gender")
}
\references{
Moody, Jim (2001) "Race, school integration, and friendship segregation in
America", American Journal of Sociology, 107(3):679--377
}
\seealso{
Other segregation measures: 
\code{\link{assort}()},
\code{\link{coleman}()},
\code{\link{ei}()},
\code{\link{gamix}()},
\code{\link{ssi}()}
}
\concept{segregation measures}
