% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classroom.R
\docType{data}
\name{Classroom}
\alias{Classroom}
\title{Classroom network}
\format{Object of class "igraph" with a directed network of size 26. Vertex
attribute \code{gender}, takes values "Boy" or "Girl".}
\usage{
Classroom
}
\description{
A directed network of ties "with whom do you like to play with?" collected in
a school class of 9-year olds. The data comes from a study undertook by
Educational Research Institute (Dolata 2014).
}
\references{
Dolata, Roman (ed). (2014). \emph{Czy szkoła ma znaczenie? Zróżnicowanie wyników
nauczania po pierwszym etapie edukacyjnym oraz jego pozaszkolne i szkolne
uwarunkowania}. Vol. 1.  Warsaw: Instytut Badań Edukacyjnych.
}
\keyword{datasets}
