% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleman.R
\name{coleman}
\alias{coleman}
\alias{coleman.table}
\alias{coleman.igraph}
\alias{coleman.default}
\title{Coleman's homophily index}
\usage{
coleman(object, ...)

\method{coleman}{table}(object, gsizes = NULL, loops = FALSE, ...)

\method{coleman}{igraph}(object, vattr, ...)

\method{coleman}{default}(object, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from methods}

\item{gsizes}{numeric vector of group sizes}

\item{loops}{logical, whether loops are allowed}

\item{vattr}{character, vertex attribute}
}
\value{
Vector of numeric values of the index for each group
}
\description{
Colemans's homphily index for directed networks.
}
\details{
Coleman's homophily index computes homophily scores for each group
defined by a vertex attribute.

If \code{object} is a table it is interpreted as a mixing matrix. If it is
only the contact layer (2-dimensional), then vector of group sizes need to
be supplied via \code{gsizes}.

\code{object} can be of class "igraph"

Default method tries to coerce \code{object} to table and use other methods.
}
\examples{
if( require(igraph, quietly = TRUE)) {
# Converting network to directed
coleman(as.directed(WhiteKinship, "mutual"), "gender")
coleman(as.directed(EF3, "mutual"), "race")
}
}
\references{
Coleman, J. (1958) "Relational analysis: The study of social organizations
with survey methods", Human Organization 17:28--36.
}
\seealso{
Other segregation measures: 
\code{\link{assort}()},
\code{\link{ei}()},
\code{\link{gamix}()},
\code{\link{orwg}()},
\code{\link{smi}()},
\code{\link{ssi}()}
}
\concept{segregation measures}
