% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei.R
\name{ei}
\alias{ei}
\alias{ei.table}
\alias{ei.igraph}
\alias{ei.default}
\title{Krackhard and Stern's E-I index}
\usage{
ei(object, ...)

\method{ei}{table}(object, ...)

\method{ei}{igraph}(
  object,
  vattr,
  directed = is.directed(object),
  loops = any(is.loop(object)),
  ...
)

\method{ei}{default}(object, ...)
}
\arguments{
\item{object}{R object, see Details for available methods}

\item{...}{other arguments passed to/from other methods}

\item{vattr}{character scalar or vector of length equal to the size of
\code{object}, vertex attribute for which mixing matrix is to be computed}

\item{directed}{logical whether the network is directed}

\item{loops}{logical, whether loops are allowed}
}
\value{
Numerical value of the E-I index.
}
\description{
An index proposed by Krackhard and Stern (1988) to capture relative
prevalence of between- and within-group ties. From that perspective it can
be interpreted as a measure of network segregation.
}
\details{
If \code{object} is a table, it is assumed to be a mixing matrix.

Method for igraphs

Default method tries to coerce \code{object} to table.
}
\examples{
ei(WhiteKinship, "gender")
}
\references{
Krackhardt, D., & Stern, R. N. (1988). Informal networks and
organizational crises: An experimental simulation. Social psychology
quarterly, 123-140.
}
\seealso{
Other segregation measures: 
\code{\link{assort}()},
\code{\link{coleman}()},
\code{\link{gamix}()},
\code{\link{ssi}()}
}
\concept{segregation measures}
